requirements.txt
streamlit
python-dotenv
psycopg2-binary
beautifulsoup4
tqdm
google-generativeai
requests
sqlalchemy